requirements.txt
streamlit>=1.28.0
python-magic>=0.4.27
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=3.0.0
docx2txt>=0.8